test: speed up Vitest API tests
