+++ frontend/src/app/api/auth/accounts/[accountId]/caregiver-capability/route.ts
@@ -49,6 +49,10 @@ async function proxyCaregiverCapability(
     }
   }
 
+  if (response.status === 204) {
+    return new NextResponse(null, { status: 204 });
+  }
+
   const text = await response.text();
   return new NextResponse(text, {
     status: response.status,
